requirements.txt
requests
pandas
numpy
scikit-learn
matplotlib
python-dotenv